+++ apps/docs/src/app/core/component-docs/checkbox/examples/checkbox-custom-values-example.component.ts
@@ -17,6 +17,7 @@ import { Component } from '@angular/core';
             <fd-checkbox
                 label="Banana is the best fruit"
                 [tristate]="true"
+                [tristateSelectable]="true"
                 [values]="{ trueValue: 'Yes', falseValue: 'No', thirdStateValue: 'I dont have an opinion' }"
                 [(ngModel)]="checkboxValue2"
             >