+++ libs/core/src/lib/message-box/message-box-header/message-box-header.component.ts
@@ -1,10 +1,16 @@
-import { AfterContentInit, ChangeDetectionStrategy, ChangeDetectorRef, Component, Optional } from '@angular/core';
-import { DialogHeaderBase } from '@fundamental-ngx/core/dialog';
 import {
-    MESSAGE_BOX_CONFIGURABLE_ELEMENT,
-    MessageBoxConfig,
-    MessageBoxConfigurableElement
-} from '../utils/message-box-config.class';
+    AfterViewInit,
+    ChangeDetectionStrategy,
+    ChangeDetectorRef,
+    Component,
+    ContentChild,
+    ElementRef,
+    Optional,
+    ViewChild
+} from '@angular/core';
+import { DialogHeaderBase } from '@fundamental-ngx/core/dialog';
+import { MessageBoxSemanticIconComponent } from '../message-box-semantic-icon/message-box-semantic-icon.component';
+import { MessageBoxConfig, MessageBoxHost } from '../utils/message-box-config.class';
 
 /**
  * Building block of the message box used to create message box header.
@@ -25,29 +31,40 @@ import {
 @Component({
     selector: 'fd-message-box-header',
     templateUrl: './message-box-header.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    providers: [{ provide: MESSAGE_BOX_CONFIGURABLE_ELEMENT, useExisting: MessageBoxHeaderComponent, multi: true }]
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class MessageBoxHeaderComponent
-    extends DialogHeaderBase
-    implements AfterContentInit, MessageBoxConfigurableElement
-{
+export class MessageBoxHeaderComponent extends DialogHeaderBase implements AfterViewInit {
     /** @hidden */
-    constructor(@Optional() public messageBoxConfig: MessageBoxConfig, changeDetectorRef: ChangeDetectorRef) {
-        super(changeDetectorRef);
-        this.messageBoxConfig = this.messageBoxConfig || {};
+    @ViewChild('defaultTemplateHeaderBar')
+    defaultTemplateHeaderBar?: ElementRef<Element>;
+
+    /** @hidden */
+    @ContentChild(MessageBoxSemanticIconComponent)
+    _projectedSemanticIcon: MessageBoxSemanticIconComponent;
+
+    /** @hidden */
+    get messageBoxConfig(): MessageBoxConfig {
+        return this.messageBox?._messageBoxConfig || {};
+    }
+
+    /** @hidden */
+    constructor(private _cdr: ChangeDetectorRef, @Optional() private messageBox?: MessageBoxHost) {
+        super(_cdr);
     }
 
     /** @hidden */
     get _showSemanticIcon(): boolean {
         return (
-            (this.messageBoxConfig.type && this.messageBoxConfig.showSemanticIcon) ||
-            !!this.messageBoxConfig.customSemanticIcon
+            !this._projectedSemanticIcon &&
+            ((this.messageBoxConfig.type && this.messageBoxConfig.showSemanticIcon) ||
+                !!this.messageBoxConfig.customSemanticIcon)
         );
     }
 
     /** @hidden */
-    ngAfterContentInit(): void {
-        super.ngAfterContentInit();
+    ngAfterViewInit(): void {
+        // need to re-check bindings after initialisation in order to reflect the proper state of "_showSemanticIcon" getter
+        // running this in next microtask to avoid "ExpressionChangedAfterItHasBeenCheckedError"
+        Promise.resolve().then(() => this._cdr.markForCheck());
     }
 }